+++ contracts/finality/src/bin/schema.rs
@@ -1,12 +1,11 @@
 use cosmwasm_schema::write_api;
-use cosmwasm_std::Empty;
-use finality::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use finality::msg::{ExecuteMsg, InstantiateMsg, MigrateMsg, QueryMsg};
 
 fn main() {
     write_api! {
         instantiate: InstantiateMsg,
         query: QueryMsg,
-        migrate: Empty,
+        migrate: MigrateMsg,
         execute: ExecuteMsg,
     }
 }